requirements.txt
asgiref==3.5.2
django-filter==21.1
Django==4.0.5
djangorestframework==3.13.1
pip==21.3.1
pytz==2022.1
setuptools==60.2.0
sqlparse==0.4.2
tzdata==2022.1
wheel==0.37.1
